pyproject.toml
[tool.poetry]
name = "lfp-image-preprocessor"
version = "0.1.0"
description = "Image tiling & tag-search-precomputer for the LeftEyePro.com website"
authors = ["c7c8 <me@crmyers.dev>"]
license = "GPLv3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.2"
pillow = "^10.4.0"
defusedxml = "^0.7.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
